Made simulated data generation even better

+++ src/kale.py
@@ -263,7 +263,7 @@ def run_tf_analysis(
             # data_for_processing = raw_gene_exp.rank(axis=0, method='average', ascending=False)
 
             # Ignore zeros
-            data_for_processing = raw_gene_exp.copy()
+            data_for_processing = raw_gene_exp.astype('float64').copy()
             mask = raw_gene_exp != 0
             data_for_processing[mask] = raw_gene_exp[mask].rank(axis=0, method='average', ascending=True)
             data_for_processing[~mask] = 0
